refactor: change temperature type to double and update API method names for weather and news fetching

+++ android/app/src/main/java/com/bilkom/model/WeatherForecast.java
@@ -16,7 +16,7 @@ import com.google.gson.annotations.SerializedName;
  */
 public class WeatherForecast {
     @SerializedName("temperature")
-    private float temperature;
+    private double temperature;
     
     @SerializedName("description")
     private String description;
